requirements.txt
opencv-python
numpy
pillow
PySide6
